requirements.txt
langchain-openai==0.3.23
langchain==0.3.25
python-dotenv